Renaming types and starting conversions
