+++ .github/workflows/nav2-integration-tests.yaml
@@ -1,4 +1,4 @@
-name: integration-tests
+name: nav2-integration-tests
 
 on: push
 
@@ -22,7 +22,7 @@ jobs:
           sudo apt update && sudo apt install docker-compose -y
 
       - name: Start test fixture containers
-        run: docker-compose -f ".github/docker/integration-tests/docker-compose-nav2.yaml" up -d --build
+        run: docker-compose -f ".github/docker/integration-tests/nav2-docker-compose.yaml" up -d --build
 
       - name: build-and-test
         uses: ./.github/actions/build-and-test
@@ -33,7 +33,7 @@ jobs:
 
       - name: Stop test fixture containers
         if: always()
-        run: docker-compose -f ".github/docker/integration-tests/docker-compose-nav2.yaml" down
+        run: docker-compose -f ".github/docker/integration-tests/nav2-docker-compose.yaml" down
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
