+++ .github/workflows/nightly.yaml
@@ -1,12 +1,14 @@
 name: nightly
 
-on:
-  schedule:
-    # 2am SGT
-    - cron: '0 18 * * *'
+# on:
+#   schedule:
+#     # 2am SGT
+#     - cron: '0 18 * * *'
+
+on: [push]
 
 jobs:
-  build-minimal-docker-images:
+  build-minimal-nav2-docker-images:
     name: Push minimal docker images to GitHub Packages
     runs-on: ubuntu-latest
     strategy:
@@ -45,12 +47,31 @@ jobs:
           tags: ghcr.io/${{ github.repository }}/minimal-zenoh-bridge-ros2dds:${{ matrix.ros_distribution }}-latest
           context: .github/docker/minimal-zenoh-bridge-ros2dds
 
+  build-minimal-nav1-docker-images:
+    name: Push minimal docker images to GitHub Packages
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        ros_distribution: [noetic]
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Login to docker
+        uses: docker/login-action@v3
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
       - name: Build and push minimal-ros1-sim
         uses: docker/build-push-action@v5
         with:
           push: true
           build-args: |
-            ROS_DISTRO=noetic
+            ROS_DISTRO=${{ matrix.ros_distribution }}
           tags: ghcr.io/${{ github.repository }}/minimal-ros1-sim:${{ matrix.ros_distribution }}-latest
           context: .github/docker/minimal-ros1-sim
 
@@ -59,7 +80,7 @@ jobs:
         with:
           push: true
           build-args: |
-            ROS_DISTRO=noetic
+            ROS_DISTRO=${{ matrix.ros_distribution }}
           tags: ghcr.io/${{ github.repository }}/minimal-nav1-bringup:${{ matrix.ros_distribution }}-latest
           context: .github/docker/minimal-nav1-bringup
 
@@ -68,14 +89,14 @@ jobs:
         with:
           push: true
           build-args: |
-            ROS_DISTRO=noetic
+            ROS_DISTRO=${{ matrix.ros_distribution }}
             ZENOH_VERSION=1.0.0-beta.2
-            FREE_FLEET_BRANCH=easy-full-control
+            FREE_FLEET_BRANCH=efc/ros1
           tags: ghcr.io/${{ github.repository }}/minimal-zenoh-bridge-ros1:${{ matrix.ros_distribution }}-latest
           context: .github/docker/minimal-zenoh-bridge-ros1
 
-  integration-tests:
-    needs: build-minimal-docker-images
+  nav2-integration-tests:
+    needs: build-minimal-nav2-docker-images
     timeout-minutes: 10
     runs-on: ubuntu-latest
     container:
@@ -94,7 +115,7 @@ jobs:
           sudo apt update && sudo apt install docker-compose -y
 
       - name: Start test fixture containers
-        run: docker-compose -f ".github/docker/integration-tests/docker-compose-nav2.yaml" up -d --build
+        run: docker-compose -f ".github/docker/integration-tests/nav2-docker-compose.yaml" up -d --build
 
       - name: build-and-test
         uses: ./.github/actions/build-and-test
@@ -105,4 +126,4 @@ jobs:
 
       - name: Stop test fixture containers
         if: always()
-        run: docker-compose -f ".github/docker/integration-tests/docker-compose-nav2.yaml" down
+        run: docker-compose -f ".github/docker/integration-tests/nav2-docker-compose.yaml" down
