+++ free_fleet_examples/config/zenoh/turtlebot3_1_zenoh_bridge_ros2dds_client_config.json5
@@ -13,15 +13,12 @@
       pub_max_frequencies: [".*/navigate_to_pose/*=1", ".*/battery_state=1"],
     },
   },
+  // Zenoh related configurations
   mode: "client",
   connect: {
-    endpoints: [
-      // "<proto>/<ip>:<port>"
-    ]
+    endpoints: []
   },
   listen: {
-    endpoints: [
-      // "<proto>/<ip>:<port>"
-    ]
+    endpoints: []
   },
 }
