+++ free_fleet_examples/free_fleet_examples/nav1_send_move_base_simple_goal.py
@@ -1,0 +1,2 @@
+# rostopic pub /move_base_simple/goal geometry_msgs/PoseStamped '{header: {stamp: now, frame_id: "map"}, pose: {position: {x: 1.0, y: 0.0, z: 0.0}, orientation: {w: 1.0}}}'
+# rostopic echo /move_base/status
